requirements.txt
opengradient>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
web3>=6.0.0
pydantic>=2.0.0
httpx>=0.27.0
tenacity>=8.2.0
requests>=2.31.0
numpy>=1.26.0
langchain>=0.2.0
langchain-community>=0.2.0
eth-account>=0.11.0
